Extract DefraKms unit tests to sibling file
Move the inline test module from defra_kms.rs into defra_kms_tests.rs
via #[path], preserving module paths, fixtures, and vectors.
